requirements.txt
fastapi
uvicorn
pytesseract
Pillow
confluent-kafka
pymongo
python-dotenv
pathlib
